requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
Werkzeug==3.1.3
gunicorn==23.0.0
SQLAlchemy==2.0.43
Jinja2==3.1.6
pytz==2025.2
python-dotenv==1.1.1
requests==2.32.4
pandas==2.3.1
openpyxl==3.2.0b1
psycopg2-binary==2.9.10
